refactor:  코드 컨벤션에 맞게 label service 및 dto 리팩토링
