+++ be/src/main/java/com/codesquad/issuetracker/api/label/dto/response/LabelResponse.java
@@ -1,4 +1,4 @@
-package com.codesquad.issuetracker.api.label.dto;
+package com.codesquad.issuetracker.api.label.dto.response;
 
 import com.codesquad.issuetracker.api.label.domain.Label;
 import lombok.Builder;
@@ -16,11 +16,11 @@ public class LabelResponse {
 
     public static LabelResponse from(Label label) {
         return LabelResponse.builder()
-            .id(label.getId())
-            .title(label.getTitle())
-            .description(label.getDescription())
-            .backgroundColor(label.getBackgroundColor())
-            .isDark(label.getIsDark())
-            .build();
+                .id(label.getId())
+                .title(label.getTitle())
+                .description(label.getDescription())
+                .backgroundColor(label.getBackgroundColor())
+                .isDark(label.getIsDark())
+                .build();
     }
 }
