+++ be/src/main/java/com/codesquad/issuetracker/api/label/service/LabelService.java
@@ -1,15 +1,16 @@
 package com.codesquad.issuetracker.api.label.service;
 
 import com.codesquad.issuetracker.api.label.domain.Label;
-import com.codesquad.issuetracker.api.label.dto.LabelCreateRequest;
-import com.codesquad.issuetracker.api.label.dto.LabelResponse;
-import com.codesquad.issuetracker.api.label.dto.LabelUpdateRequest;
+import com.codesquad.issuetracker.api.label.dto.request.LabelCreateRequest;
+import com.codesquad.issuetracker.api.label.dto.request.LabelUpdateRequest;
+import com.codesquad.issuetracker.api.label.dto.response.LabelResponse;
 import com.codesquad.issuetracker.api.label.repository.LabelRepository;
 import com.codesquad.issuetracker.api.organization.repository.OrganizationRepository;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -18,23 +19,26 @@ public class LabelService {
     private final LabelRepository labelRepository;
     private final OrganizationRepository organizationRepository;
 
-    public List<LabelResponse> readAll(String organizationTitle) {
-        Long organizationId = organizationRepository.findBy(organizationTitle).orElseThrow();
-        return labelRepository.findAllBy(organizationId).stream()
-            .map(LabelResponse::from)
-            .collect(Collectors.toUnmodifiableList());
-    }
-
+    @Transactional
     public Long create(String organizationTitle, LabelCreateRequest labelCreateRequest) {
         Long organizationId = organizationRepository.findBy(organizationTitle).orElseThrow();
-        Label label = LabelCreateRequest.toEntity(organizationId, labelCreateRequest);
+        Label label = labelCreateRequest.toEntity(organizationId);
         return labelRepository.save(label).orElseThrow();
     }
 
-    public Long update(String organizationTitle, LabelUpdateRequest labelUpdateRequest,
-        Long labelId) {
+    @Transactional
+    public List<LabelResponse> readAll(String organizationTitle) {
         Long organizationId = organizationRepository.findBy(organizationTitle).orElseThrow();
-        Label label = LabelUpdateRequest.toEntity(labelUpdateRequest, organizationId, labelId);
+        return labelRepository.findAllBy(organizationId).stream()
+                .map(LabelResponse::from)
+                .collect(Collectors.toUnmodifiableList());
+    }
+
+    @Transactional
+    public Long update(String organizationTitle, LabelUpdateRequest labelUpdateRequest,
+                       Long labelId) {
+        Long organizationId = organizationRepository.findBy(organizationTitle).orElseThrow();
+        Label label = labelUpdateRequest.toEntity(organizationId, labelId);
         return labelRepository.update(label);
     }
